make operation sync cover all protocol
