+++ server/src/main/java/org/apache/iotdb/db/service/basic/ServiceProvider.java
@@ -59,7 +59,7 @@ import static org.apache.iotdb.db.utils.ErrorHandlingUtils.onNPEOrUnexpectedExce
 
 public abstract class ServiceProvider {
 
-  private static final Logger LOGGER = LoggerFactory.getLogger(ServiceProvider.class);
+  protected static final Logger LOGGER = LoggerFactory.getLogger(ServiceProvider.class);
   public static final Logger AUDIT_LOGGER =
       LoggerFactory.getLogger(IoTDBConstant.AUDIT_LOGGER_NAME);
   public static final Logger SLOW_SQL_LOGGER =
